weather-etl-transform/requirements.txt
functions-framework==3.8.0
google-cloud-bigquery==3.25.0